+++ src/BigCommerce/ResourceModels/Theme/JobResult.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace BigCommerce\ApiV3\ResourceModels\Theme;
+
+use BigCommerce\ApiV3\ResourceModels\ResourceModel;
+
+class JobResult extends ResourceModel
+{
+    public string $theme_id;
+}