+++ RELEASE_NOTES.md
@@ -1,6 +1,13 @@
+#### New Features
 
+Add support for the [Themes API](https://developer.bigcommerce.com/api-reference/store-management/themes)
 
+Includes
 
-Code Improvements
+- Themes
+- Theme Actions
+- Theme Jobs
+
+#### Code Improvements
 
 - Refactor PaginatedResponse to be simpler to implement